LPS-129897 Create new  item selector criteria based in the role type to allow select more than one type per selector window (and deprecate the old criterion one)
